Cleaned

+++ utils.py
@@ -1,21 +1,13 @@
-import torch
-import pickle
-import PIL.Image
-from multiprocessing.pool import ThreadPool
-import torch.nn.functional as F
-import functools
-import numpy as np
-import pickle
 import os
+
+import PIL.Image
 import cv2
-import math
-import matplotlib.pyplot as plt
-import torch.nn as nn
-import torch.distributions
-import torchvision.transforms as transforms
-from PIL import Image
 import dominate
+import numpy as np
+import torch
+import torch.distributions
 from dominate.tags import *
+
 
 # os.environ['CUDA_VISIBLE_DEVICES'] = '0'
 # device = torch.device("cuda:0" if (torch.cuda.is_available()) else "cpu")
@@ -34,9 +26,9 @@ def load_img(path, target_size):
             img = img.convert('RGB')
     wh_tuple = (target_size[1], target_size[0])
     if img.size != wh_tuple:
-        img = img.resize(wh_tuple, resample = PIL.Image.BILINEAR)
+        img = img.resize(wh_tuple, resample=PIL.Image.BILINEAR)
 
-    x = np.asarray(img, dtype = "uint8")
+    x = np.asarray(img, dtype="uint8")
     if len(x.shape) == 2:
         x = np.expand_dims(x, -1)
 
@@ -58,17 +50,18 @@ def postprocess(x):
 
 def tile(X, rows, cols):
     """Tile images for display."""
-    tiling = np.zeros((rows * X.shape[1], cols * X.shape[2], X.shape[3]), dtype = X.dtype)
+    tiling = np.zeros((rows * X.shape[1], cols * X.shape[2], X.shape[3]), dtype=X.dtype)
     for i in range(rows):
         for j in range(cols):
             idx = i * cols + j
             if idx < X.shape[0]:
-                img = X[idx,...]
+                img = X[idx, ...]
                 tiling[
-                        i*X.shape[1]:(i+1)*X.shape[1],
-                        j*X.shape[2]:(j+1)*X.shape[2],
-                        :] = img
+                i * X.shape[1]:(i + 1) * X.shape[1],
+                j * X.shape[2]:(j + 1) * X.shape[2],
+                :] = img
     return tiling
+
 
 def make_joint_img(img_shape, jo, joints):
     # three channels: left, right, center
@@ -76,62 +69,61 @@ def make_joint_img(img_shape, jo, joints):
     thickness = int(3 * scale_factor)
     imgs = list()
     for i in range(3):
-        imgs.append(np.zeros(img_shape[:2], dtype = "uint8"))
+        imgs.append(np.zeros(img_shape[:2], dtype="uint8"))
 
     body = ["lhip", "lshoulder", "rshoulder", "rhip"]
-    body_pts = np.array([[joints[jo.index(part),:] for part in body]])
+    body_pts = np.array([[joints[jo.index(part), :] for part in body]])
     if np.min(body_pts) >= 0:
         body_pts = np.int_(body_pts)
         cv2.fillPoly(imgs[2], body_pts, 255)
 
     right_lines = [
-            ("rankle", "rknee"),
-            ("rknee", "rhip"),
-            ("rhip", "rshoulder"),
-            ("rshoulder", "relbow"),
-            ("relbow", "rwrist")]
+        ("rankle", "rknee"),
+        ("rknee", "rhip"),
+        ("rhip", "rshoulder"),
+        ("rshoulder", "relbow"),
+        ("relbow", "rwrist")]
     for line in right_lines:
         l = [jo.index(line[0]), jo.index(line[1])]
         if np.min(joints[l]) >= 0:
             a = tuple(np.int_(joints[l[0]]))
             b = tuple(np.int_(joints[l[1]]))
-            cv2.line(imgs[0], a, b, color = 255, thickness = thickness)
+            cv2.line(imgs[0], a, b, color=255, thickness=thickness)
 
     left_lines = [
-            ("lankle", "lknee"),
-            ("lknee", "lhip"),
-            ("lhip", "lshoulder"),
-            ("lshoulder", "lelbow"),
-            ("lelbow", "lwrist")]
+        ("lankle", "lknee"),
+        ("lknee", "lhip"),
+        ("lhip", "lshoulder"),
+        ("lshoulder", "lelbow"),
+        ("lelbow", "lwrist")]
     for line in left_lines:
         l = [jo.index(line[0]), jo.index(line[1])]
         if np.min(joints[l]) >= 0:
             a = tuple(np.int_(joints[l[0]]))
             b = tuple(np.int_(joints[l[1]]))
-            cv2.line(imgs[1], a, b, color = 255, thickness = thickness)
+            cv2.line(imgs[1], a, b, color=255, thickness=thickness)
 
     rs = joints[jo.index("rshoulder")]
     ls = joints[jo.index("lshoulder")]
     cn = joints[jo.index("cnose")]
-    neck = 0.5*(rs+ls)
+    neck = 0.5 * (rs + ls)
     a = tuple(np.int_(neck))
     b = tuple(np.int_(cn))
     if np.min(a) >= 0 and np.min(b) >= 0:
-        cv2.line(imgs[0], a, b, color = 127, thickness = thickness)
-        cv2.line(imgs[1], a, b, color = 127, thickness = thickness)
+        cv2.line(imgs[0], a, b, color=127, thickness=thickness)
+        cv2.line(imgs[1], a, b, color=127, thickness=thickness)
 
     cn = tuple(np.int_(cn))
     leye = tuple(np.int_(joints[jo.index("leye")]))
     reye = tuple(np.int_(joints[jo.index("reye")]))
     if np.min(reye) >= 0 and np.min(leye) >= 0 and np.min(cn) >= 0:
-        cv2.line(imgs[0], cn, reye, color = 255, thickness = thickness)
-        cv2.line(imgs[1], cn, leye, color = 255, thickness = thickness)
+        cv2.line(imgs[0], cn, reye, color=255, thickness=thickness)
+        cv2.line(imgs[1], cn, leye, color=255, thickness=thickness)
 
-    img = np.stack(imgs, axis = -1)
+    img = np.stack(imgs, axis=-1)
     if img_shape[-1] == 1:
-        img = np.mean(img, axis = -1)[:,:,None]
+        img = np.mean(img, axis=-1)[:, :, None]
     return img
-
 
 
 def valid_joints(*joints):
@@ -139,7 +131,7 @@ def valid_joints(*joints):
     return (j >= 0).all()
 
 
-def get_crop(bpart, joints, jo, wh, o_w, o_h, ar = 1.0):
+def get_crop(bpart, joints, jo, wh, o_w, o_h, ar=1.0):
     bpart_indices = [jo.index(b) for b in bpart]
     part_src = np.float32(joints[bpart_indices])
 
@@ -149,24 +141,23 @@ def get_crop(bpart, joints, jo, wh, o_w, o_h, ar = 1.0):
             bpart = ["lhip"]
             bpart_indices = [jo.index(b) for b in bpart]
             part_src = np.float32(joints[bpart_indices])
-        elif bpart[0] == "rhip" and bpart[1] == "rknee": 
+        elif bpart[0] == "rhip" and bpart[1] == "rknee":
             bpart = ["rhip"]
             bpart_indices = [jo.index(b) for b in bpart]
             part_src = np.float32(joints[bpart_indices])
-        elif bpart[0] == "lshoulder" and bpart[1] == "rshoulder" and bpart[2] == "cnose": 
+        elif bpart[0] == "lshoulder" and bpart[1] == "rshoulder" and bpart[2] == "cnose":
             bpart = ["lshoulder", "rshoulder", "rshoulder"]
             bpart_indices = [jo.index(b) for b in bpart]
             part_src = np.float32(joints[bpart_indices])
 
-
     if not valid_joints(part_src):
-            return None
+        return None
 
     if part_src.shape[0] == 1:
         # leg fallback
         a = part_src[0]
-        b = np.float32([a[0],o_h - 1])
-        part_src = np.float32([a,b])
+        b = np.float32([a[0], o_h - 1])
+        part_src = np.float32([a, b])
 
     if part_src.shape[0] == 4:
         pass
@@ -174,7 +165,7 @@ def get_crop(bpart, joints, jo, wh, o_w, o_h, ar = 1.0):
         # lshoulder, rshoulder, cnose
         if bpart == ["lshoulder", "rshoulder", "rshoulder"]:
             segment = part_src[1] - part_src[0]
-            normal = np.array([-segment[1],segment[0]])
+            normal = np.array([-segment[1], segment[0]])
             if normal[1] > 0.0:
                 normal = -normal
 
@@ -182,36 +173,36 @@ def get_crop(bpart, joints, jo, wh, o_w, o_h, ar = 1.0):
             b = part_src[0]
             c = part_src[1]
             d = part_src[1] + normal
-            part_src = np.float32([a,b,c,d])
+            part_src = np.float32([a, b, c, d])
         else:
             assert bpart == ["lshoulder", "rshoulder", "cnose"]
-            neck = 0.5*(part_src[0] + part_src[1])
+            neck = 0.5 * (part_src[0] + part_src[1])
             neck_to_nose = part_src[2] - neck
-            part_src = np.float32([neck + 2*neck_to_nose, neck])
+            part_src = np.float32([neck + 2 * neck_to_nose, neck])
 
             # segment box
             segment = part_src[1] - part_src[0]
-            normal = np.array([-segment[1],segment[0]])
+            normal = np.array([-segment[1], segment[0]])
             alpha = 1.0 / 2.0
-            a = part_src[0] + alpha*normal
-            b = part_src[0] - alpha*normal
-            c = part_src[1] - alpha*normal
-            d = part_src[1] + alpha*normal
-            #part_src = np.float32([a,b,c,d])
-            part_src = np.float32([b,c,d,a])
+            a = part_src[0] + alpha * normal
+            b = part_src[0] - alpha * normal
+            c = part_src[1] - alpha * normal
+            d = part_src[1] + alpha * normal
+            # part_src = np.float32([a,b,c,d])
+            part_src = np.float32([b, c, d, a])
     else:
         assert part_src.shape[0] == 2
 
         segment = part_src[1] - part_src[0]
-        normal = np.array([-segment[1],segment[0]])
+        normal = np.array([-segment[1], segment[0]])
         alpha = ar / 2.0
-        a = part_src[0] + alpha*normal
-        b = part_src[0] - alpha*normal
-        c = part_src[1] - alpha*normal
-        d = part_src[1] + alpha*normal
-        part_src = np.float32([a,b,c,d])
+        a = part_src[0] + alpha * normal
+        b = part_src[0] - alpha * normal
+        c = part_src[1] - alpha * normal
+        d = part_src[1] + alpha * normal
+        part_src = np.float32([a, b, c, d])
 
-    dst = np.float32([[0.0,0.0],[0.0,1.0],[1.0,1.0],[1.0,0.0]])
+    dst = np.float32([[0.0, 0.0], [0.0, 1.0], [1.0, 1.0], [1.0, 0.0]])
     part_dst = np.float32(wh * dst)
 
     M = cv2.getPerspectiveTransform(part_src, part_dst)
@@ -228,48 +219,49 @@ def normalize(imgs, coords, stickmen, jo, box_factor):
         joints = coords[i]
         stickman = stickmen[i]
 
-        h,w = img.shape[:2]
-#         print("h",h,"w",w)
+        h, w = img.shape[:2]
+        #         print("h",h,"w",w)
         o_h = h
         o_w = w
-        h = h // 2**box_factor
-        w = w // 2**box_factor
-        wh = np.array([w,h])
+        h = h // 2 ** box_factor
+        w = w // 2 ** box_factor
+        wh = np.array([w, h])
         wh = np.expand_dims(wh, 0)
 
         bparts = [
-                ["lshoulder","lhip","rhip","rshoulder"],
-                ["lshoulder", "rshoulder", "cnose"],
-                ["lshoulder","lelbow"],
-                ["lelbow", "lwrist"],
-                ["rshoulder","relbow"],
-                ["relbow", "rwrist"],
-                ["lhip", "lknee"],
-                ["rhip", "rknee"]]
+            ["lshoulder", "lhip", "rhip", "rshoulder"],
+            ["lshoulder", "rshoulder", "cnose"],
+            ["lshoulder", "lelbow"],
+            ["lelbow", "lwrist"],
+            ["rshoulder", "relbow"],
+            ["relbow", "rwrist"],
+            ["lhip", "lknee"],
+            ["rhip", "rknee"]]
         ar = 0.5
 
         part_imgs = list()
         part_stickmen = list()
         for bpart in bparts:
-            part_img = np.zeros((h,w,3))
-            part_stickman = np.zeros((h,w,3))
+            part_img = np.zeros((h, w, 3))
+            part_stickman = np.zeros((h, w, 3))
             M = get_crop(bpart, joints, jo, wh, o_w, o_h, ar)
-#             print(M)
+            #             print(M)
 
             if M is not None:
-                part_img = cv2.warpPerspective(img, M, (h,w), borderMode = cv2.BORDER_REPLICATE)
-                part_stickman = cv2.warpPerspective(stickman, M, (h,w), borderMode = cv2.BORDER_REPLICATE)
+                part_img = cv2.warpPerspective(img, M, (h, w), borderMode=cv2.BORDER_REPLICATE)
+                part_stickman = cv2.warpPerspective(stickman, M, (h, w), borderMode=cv2.BORDER_REPLICATE)
 
             part_imgs.append(part_img)
             part_stickmen.append(part_stickman)
-        img = np.concatenate(part_imgs, axis = 2)
-        stickman = np.concatenate(part_stickmen, axis = 2)
+        img = np.concatenate(part_imgs, axis=2)
+        stickman = np.concatenate(part_stickmen, axis=2)
 
         out_imgs.append(img)
         out_stickmen.append(stickman)
     out_imgs = np.stack(out_imgs)
     out_stickmen = np.stack(out_stickmen)
     return out_imgs, out_stickmen
+
 
 class HTML:
     def __init__(self, web_dir, title, reflesh=0):
@@ -315,12 +307,12 @@ class HTML:
         f = open(html_file, 'wt')
         f.write(self.doc.render())
         f.close()
-        
-        
+
+
 # def load_network_weights(net, path):
 #     state_dict = torch.load(path, map_location=str(device))
 #     net.load_state_dict(state_dict)
-    
+
 def init_weights(net, init_type='kaiming', gain=0.02):
     def init_func(m):
         classname = m.__class__.__name__
@@ -343,7 +335,8 @@ def init_weights(net, init_type='kaiming', gain=0.02):
 
     print('initialize network with %s' % init_type)
     net.apply(init_func)
-    
+
+
 def tensor2im(input_image, imtype=np.uint8):
     if isinstance(input_image, torch.Tensor):
         image_tensor = input_image.data
@@ -352,5 +345,5 @@ def tensor2im(input_image, imtype=np.uint8):
     image_numpy = image_tensor[0].cpu().float().numpy()
     if image_numpy.shape[0] == 1:
         image_numpy = np.tile(image_numpy, (3, 1, 1))
-    image_numpy = (np.transpose(image_numpy, (1, 2, 0))+1) / 2.0 * 255.0
+    image_numpy = (np.transpose(image_numpy, (1, 2, 0)) + 1) / 2.0 * 255.0
     return image_numpy.astype(imtype)
